+++ core_client/base/api/v3_cluster_get_snapshot.py
@@ -9,8 +9,6 @@ from ..models import Error
 @validate_call()
 def _build_request(
     client: Client,
-    id: str,
-    input_id: str,
     retries: int = None,
     timeout: float = None,
 ):
@@ -19,8 +17,8 @@ def _build_request(
     if not timeout:
         timeout = client.timeout
     return {
-        "method": "put",
-        "url": f"{client.base_url}/api/v3/process/{id}/playout/{input_id}/stream",
+        "method": "get",
+        "url": f"{client.base_url}/api/v3/cluster/snapshot",
         "headers": client.headers,
         "timeout": timeout,
         "data": None,
@@ -30,7 +28,7 @@ def _build_request(
 
 def _build_response(response: httpx.Response):
     if response.status_code == 200:
-        response_200 = response.json()
+        response_200 = response.content
         return response_200
     else:
         response_error = TypeAdapter(Error).validate_python(response.json())