+++ core_client/base/api/v3_fs_delete_file_list.py
@@ -9,8 +9,12 @@ from ..models import Error
 @validate_call()
 def _build_request(
     client: Client,
-    id: str,
-    input_id: str,
+    storage: str,
+    glob: str = "",
+    size_min: str = "",
+    size_max: str = "",
+    lastmod_start: str = "",
+    lastmod_end: str = "",
     retries: int = None,
     timeout: float = None,
 ):
@@ -19,8 +23,10 @@ def _build_request(
     if not timeout:
         timeout = client.timeout
     return {
-        "method": "put",
-        "url": f"{client.base_url}/api/v3/process/{id}/playout/{input_id}/stream",
+        "method": "delete",
+        "url": f"{client.base_url}/api/v3/fs/{storage}"
+        + f"?glob={glob}&size_min={size_min}&size_max={size_max}"
+        + f"&lastmod_start={lastmod_start}&lastmod_end={lastmod_end}",
         "headers": client.headers,
         "timeout": timeout,
         "data": None,
@@ -30,7 +36,7 @@ def _build_request(
 
 def _build_response(response: httpx.Response):
     if response.status_code == 200:
-        response_200 = response.json()
+        response_200 = TypeAdapter(list[str]).validate_python(response.json())
         return response_200
     else:
         response_error = TypeAdapter(Error).validate_python(response.json())