+++ core_client/base/models/v3/__init__.py
@@ -10,6 +10,17 @@ from .cluster_node import ClusterNode
 from .cluster_deployments_process import ClusterDeploymentsProcess
 from .cluster_deployments import ClusterDeployments
 
+from .cluster_kvs_value import ClusterKVSValue
+from .cluster_store_node import ClusterStoreNode
+
+from .log_event_filter import LogEventFilter
+from .event_filters import EventFilters
+from .log_event import LogEvent
+from .media_event import MediaEvent
+
+from .graph_query import GraphQuery
+from .graph_response import GraphResponse
+
 from .cluster_reallocation_node_process import ClusterReallocationNodeProcess
 from .cluster_reallocation_node import ClusterReallocationNode
 from .cluster_reallocation import ClusterReallocation