+++ core_client/base/models/v3/cluster_kvs_value.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+
+class ClusterKVSValue(BaseModel):
+    """
+    {
+        "updated_at": "string",
+        "value": "string"
+    }
+    """
+
+    updated_at: str | None = None
+    value: str | None = None