+++ core_client/base/models/v3/event_filters.py
@@ -1,0 +1,13 @@
+from pydantic import BaseModel
+
+from . import LogEventFilter
+
+
+class EventFilters(BaseModel):
+    """
+    {
+        "filters": [LogEventFilter]
+    }
+    """
+
+    filters: list[LogEventFilter] | None = None