+++ core_client/base/models/v3/media_event.py
@@ -1,0 +1,17 @@
+from pydantic import BaseModel
+
+
+class MediaEvent(BaseModel):
+    """
+    {
+        "action": "string",
+        "name": "string",
+        "names": ["string"],
+        "ts": int
+    }
+    """
+
+    action: str | None = None
+    name: str | None = None
+    names: list[str] | None = None
+    ts: int | None = None